requirements.txt
azure-mgmt-network==26.0.0
azure-identity==1.17.1
tabluate==0.9.0
azure-mgmt-compute==33.0.0
